\name{plot_str}
\alias{plot_str}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
FUNCTION TO CONVERT INDIVIDUAL ID IN OUTPUT FILE
}
\description{
internal function of MPI_structure. when temporary data files are generated with a subset of the total populations, populations are re-names with IDs from 1 to n.
This function converts populations ID back to the initial data file IDs 
}
\usage{
convert_IDs()
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{...}{
...
}
 }
\author{
Francois Besnier, Kevin Glover
}
