\name{structure_data}
\alias{structure_data}
\docType{data}
\title{
 EXAMPLE DATA FILE FOR STRUCTURE
}
\description{
contain microsatelite genotype on nine loci for 987 individuals divided in  8 populations
}
\usage{data(structure_data)}
\format{
 The format is:
 two lines per individual
 without row of marker names
 }
\examples{
data(structure_data)
}
\keyword{datasets}
